+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Implement new key `options` to request and endpoint nodes [#529](https://github.com/scanapi/scanapi/pull/552)
 
 ## [2.7.0] - 2022-06-02
 ### Deprecated